+++ tests/common/EncodeTest.php
