+++ tests/common/Objects/StringableObject.php
