+++ tests/common/Objects/TestBooleanInputTag.php
