+++ tests/common/Objects/TestNormalTag.php
