+++ tests/common/Tag/Base/BooleanInputTagTest.php
