+++ tests/common/Tag/Base/NormalTagTest.php
