+++ tests/common/Tag/Base/TagContentTraitTest.php
@@ -98,21 +98,6 @@ final class TagContentTraitTest extends TestCase
         );
     }
 
-    /**
-     * @requires PHP >= 8.0
-     */
-    public function testNamedParametersContent(): void
-    {
-        $this->assertSame(
-            '<test>123</test>',
-            TestTagContentTrait::tag()
-                ->content(content: '1')
-                ->addContent(content: '2')
-                ->addContent(content: '3')
-                ->render()
-        );
-    }
-
     public function testImmutability(): void
     {
         $tag = TestTagContentTrait::tag();
