+++ tests/common/Tag/Base/VoidTagTest.php
