+++ tests/common/Tag/CustomTagTest.php
