+++ tests/common/Tag/ScriptTest.php
