+++ tests/common/Tag/SpanTest.php
