+++ tests/common/Tag/StyleTest.php
