+++ tests/common/Widget/CheckboxListTest.php
